as68: fix sign extension when listing instruction words

+++ alcyon/as/list.c
@@ -195,7 +195,7 @@ PP(int pflag;)
 	{
 		pi = ins;
 		for (i = 0; i < (instrlen / 2); i++)	/* binary */
-			printf("%04x", *pi++);
+			printf("%04x", (uint16_t)*pi++);
 		if (instrlen & 1)
 		{
 			i++;
